backend/pyproject.toml
[project]
name = "echo-edger"
version = "0.1.0"
description = "Backend for Echo Ledger"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "faster-whisper>=1.2.1",
    "llama-cpp-python>=0.3.16",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.1",
    "rapidfuzz>=3.14.3",
    "silero-vad>=6.2.0",
    "torch>=2.10.0",
    "uvicorn>=0.41.0",
    "websockets>=16.0",
]
